Add files via upload
This repository contains a Power BI project that provides insights into HR metrics of the organization. The .pbit file includes a dashboard that displays key HR KPIs such as employee promotion, retrenchment, and detailed insights. The project is designed to help us understand our workforce and identify opportunities to improve employee engagement and retention. The dashboard is interactive and includes custom visuals that provide a unique user experience. Additionally, the repository includes supporting documentation and data sources used in the dashboard.